Lint and reduce diff
