+++ python/ray/data/_internal/table_block.py
@@ -7,7 +7,6 @@ from ray.air.constants import TENSOR_COLUMN_NAME
 from ray.data._internal.block_builder import BlockBuilder
 from ray.data._internal.numpy_support import convert_udf_returns_to_numpy, is_array_like
 from ray.data._internal.size_estimator import SizeEstimator
-from ray.data._internal.util import _lazy_import_pandas
 from ray.data.block import Block, BlockAccessor
 from ray.data.row import TableRow
