+++ python/ray/data/_internal/util.py
@@ -49,41 +49,31 @@ _EXAMPLE_SCHEME = "example"
 
 
 LazyModule = Union[None, bool, ModuleType]
-_pandas: LazyModule = None
-_pyarrow: LazyModule = None
 _pyarrow_dataset: LazyModule = None
+_pyarrow_table: LazyModule = None
 
 
-def _lazy_import_pandas() -> LazyModule:
-    global _pandas
-    if _pandas is None:
+def _lazy_import_pyarrow_table() -> LazyModule:
+    global _pyarrow_table
+    if _pyarrow_table is None:
         try:
-            import pandas as _pandas
+            from pyarrow import Table as _pyarrow_table
         except ModuleNotFoundError:
-            # If module is not found, set _pandas to False so we won't
-            # keep trying to import it on every _lazy_import_pandas() call.
-            _pandas = False
-    return _pandas
-
-
-def _lazy_import_pyarrow() -> LazyModule:
-    global _pyarrow
-    if _pyarrow is None:
-        try:
-            import pyarrow as _pyarrow
-        except ModuleNotFoundError:
-            # If module is not found, set _pandas to False so we won't
-            # keep trying to import it on every _lazy_import_pandas() call.
-            _pyarrow = False
-    return _pyarrow
+            # If module is not found, set _pyarrow to False so we won't
+            # keep trying to import it on every _lazy_import_pyarrow() call.
+            _pyarrow_table = False
+    return _pyarrow_table
 
 
 def _lazy_import_pyarrow_dataset() -> LazyModule:
     global _pyarrow_dataset
-    if not _pyarrow_dataset:
-        pyarrow = _lazy_import_pyarrow()
-        if pyarrow:
-            _pyarrow_dataset = pyarrow.dataset
+    if _pyarrow_dataset is None:
+        try:
+            from pyarrow import dataset as _pyarrow_dataset
+        except ModuleNotFoundError:
+            # If module is not found, set _pyarrow to False so we won't
+            # keep trying to import it on every _lazy_import_pyarrow() call.
+            _pyarrow_dataset = False
     return _pyarrow_dataset
